fix(desktop): show only relevant settings and recovery controls
