+++ app/client/modules/settings/routes/settings-shell.tsx
@@ -9,6 +9,10 @@ import { InstanceSettings } from "~/client/modules/settings/components/instance-
 import type { AppContext } from "~/context";
 import { getActiveSettingsScope, getSettingsScopeAvailability } from "../settings-scope";
 import { SettingsPage } from "./settings";
+import { useSystemInfo } from "~/client/hooks/use-system-info";
+import { Card } from "~/client/components/ui/card";
+import { DateTimeFormatSection } from "../components/date-time-format-section";
+import { RecoveryKeySection } from "../components/recovery-key-section";
 
 type Props = {
 	appContext: AppContext;
@@ -27,6 +31,27 @@ export function SettingsShell({
 }: Props) {
 	const search = useSearch({ from: "/(dashboard)/settings/" });
 	const permissions = usePermissions();
+	const { runtime } = useSystemInfo();
+
+	if (runtime === "desktop") {
+		const showRecoveryKey = permissions.can("recoveryKey.download");
+		const hasPassword = appContext.user?.hasPassword === true;
+		return (
+			<div className="space-y-6">
+				<Card className="p-0 gap-0">
+					<DateTimeFormatSection />
+				</Card>
+				{showRecoveryKey && (
+					<Card id="recovery-key" className="scroll-mt-20 p-0 gap-0">
+						<RecoveryKeySection
+							passwordAuthSupported={appContext.passwordAuthSupported}
+							hasPassword={hasPassword}
+						/>
+					</Card>
+				)}
+			</div>
+		);
+	}
 
 	const scopeAvailability = getSettingsScopeAvailability(permissions);
 	const activeScope = getActiveSettingsScope(search, scopeAvailability);
